feat(carousel): add tests to base and styled components
