+++ jest.config.js
@@ -14,4 +14,5 @@ module.exports = {
         'jest-watch-typeahead/filename',
         'jest-watch-typeahead/testname',
     ],
+    snapshotSerializers: ['@emotion/jest/serializer'],
 };
