+++ package.json
@@ -66,6 +66,7 @@
     "@babel/preset-env": "^7.9.5",
     "@babel/preset-react": "^7.9.4",
     "@commitlint/config-conventional": "^9.0.1",
+    "@emotion/jest": "^11.3.0",
     "@react-theming/storybook-addon": "^1.1.1",
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/git": "^9.0.0",
