+++ src/utils/index.js
@@ -2,8 +2,6 @@ export { default as checkSnapSupport } from './checkSnapSupport';
 export { default as cloneSlide } from './cloneSlide';
 export { default as renderSlides } from './renderSlides';
 export { default as animate } from './animate';
-export { default as timingFunctions } from './timingFunctions';
-export { quad, circ, makeEaseOut } from './timingFunctions';
 export { animateScroll } from './animate';
 export { default as easeInOutQuad } from './easeInOutQuad';
 export { default as infiniteScroll } from './unevenInfiniteScroll';
