--- tests/components/StylesMockBullets.css
@@ -1,31 +1,0 @@
-.containerDefault {
-    margin: 1rem;
-}
-
-.containerInfinite {
-    margin: 1rem;
-}
-
-.container {
-    margin: 1rem;
-}
-
-.moveInfinite {
-    margin: 1rem;
-}
-
-.bullet {
-    margin: 1rem;
-}
-
-.bulletInfinite {
-    margin: 1rem;
-}
-
-.isSecondary {
-    margin: 1rem;
-}
-
-.isActive {
-    margin: 1rem;
-}
