+++ tests/containers/BulletsContainer.test.js
@@ -21,20 +21,6 @@ describe('<BulletsContainer/>', () => {
         expect(tree).toMatchSnapshot();
     });
 
-    it('should render correctly with custom Props', () => {
-        const mockTheme = {
-            container: 'container',
-        };
-        const tree = shallow(
-            <BulletsContainer data-testId="test" theme={mockTheme} />
-        ).renderProp('children')({
-            ...mockContextCarousel,
-        });
-
-        expect(tree.find('Bullets').prop('data-testId')).toBe('test');
-        expect(tree.find('Bullets').prop('theme')).toEqual(mockTheme);
-    });
-
     it('should render correctly with correct Context', () => {
         const tree = shallow(<BulletsContainer />).renderProp('children')({
             ...mockContextCarousel,
